Move maps catalog files around
