+++ roles/maps-catalog.json
@@ -1,0 +1,11 @@
+{
+    "_README": [
+        "If you are installing or configuring",
+        "IIAB Maps, you should ignore this file.",
+        "See /opt/iiab/iiab/maps/generate-catalog.py",
+        "for more info."
+    ],
+    "satellite": "https://iiab.switnet.org/maps/2/s2maps-sentinel2-2023.2025-12-10.z00-z13.pmtiles",
+    "terrain": "https://iiab.switnet.org/maps/2/terrarium.2025-12-10.z00-z10.pmtiles",
+    "vector": "https://iiab.switnet.org/maps/2/openstreetmap-openmaptiles.2026-04-01.z00-z14.pmtiles"
+}
