--- roles/maps/catalog/catalog.json
@@ -1,1 +1,0 @@
-{"vector": "https://iiab.switnet.org/maps/2/openstreetmap-openmaptiles.2026-04-01.z00-z14.pmtiles", "satellite": "https://iiab.switnet.org/maps/2/s2maps-sentinel2-2023.2025-12-10.z00-z13.pmtiles", "terrain": "https://iiab.switnet.org/maps/2/terrarium.2025-12-10.z00-z10.pmtiles"}
